requirements.txt
flask
gunicorn
inference-sdk
requests
pillow